requirements.txt
pandas>=2.2.0
pyyaml>=6.0.1
jinja2>=3.1.4
nbformat>=5.10.4
matplotlib>=3.8.0
